+++ examples/ai-core/src/generate-text/openai-custom-fetch.ts
@@ -1,0 +1,27 @@
+import { createOpenAI } from '@ai-sdk/openai';
+import { generateText } from 'ai';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const openai = createOpenAI({
+  // example fetch wrapper that logs the URL:
+  fetch: async (url, options) => {
+    console.log(`Fetching ${url}`);
+    const result = await fetch(url, options);
+    console.log(`Fetched ${url}`);
+    console.log();
+    return result;
+  },
+});
+
+async function main() {
+  const result = await generateText({
+    model: openai('gpt-3.5-turbo'),
+    prompt: 'Invent a new holiday and describe its traditions.',
+  });
+
+  console.log(result.text);
+}
+
+main().catch(console.error);