+++ packages/azure/src/azure-openai-provider.ts
@@ -13,6 +13,14 @@ export interface AzureOpenAIProvider {
   /**
 Creates an Azure OpenAI chat model for text generation.
    */
+  languageModel(
+    deploymentId: string,
+    settings?: OpenAIChatSettings,
+  ): OpenAIChatLanguageModel;
+
+  /**
+Creates an Azure OpenAI chat model for text generation.
+   */
   chat(
     deploymentId: string,
     settings?: OpenAIChatSettings,
@@ -29,6 +37,12 @@ Name of the Azure OpenAI resource.
 API key for authenticating requests.
      */
   apiKey?: string;
+
+  /**
+Custom fetch implementation. You can use it as a middleware to intercept requests,
+or to provide a custom fetch implementation for e.g. testing.
+    */
+  fetch?: typeof fetch;
 }
 
 /**
@@ -63,6 +77,7 @@ export function createAzure(
       url: ({ path, modelId }) =>
         `https://${getResourceName()}.openai.azure.com/openai/deployments/${modelId}${path}?api-version=2024-05-01-preview`,
       compatibility: 'compatible',
+      fetch: options.fetch,
     });
 
   const provider = function (
@@ -78,6 +93,7 @@ export function createAzure(
     return createChatModel(deploymentId, settings as OpenAIChatSettings);
   };
 
+  provider.languageModel = createChatModel;
   provider.chat = createChatModel;
 
   return provider as AzureOpenAIProvider;